requirements.txt
typer
pydantic
pyyaml
requests
pypdf
python-docx
openpyxl